+++ lib/pbio/drv/uart/uart_stm32f4_ll_irq.c
@@ -39,15 +39,15 @@ struct _pbdrv_uart_dev_t {
     /** The buffer of the ongoing async read function. */
     uint8_t *read_buf;
     /** The length of read_buf in bytes. */
-    uint8_t read_length;
+    uint32_t read_length;
     /** The current position in read_buf. */
-    uint8_t read_pos;
+    uint32_t read_pos;
     /** The buffer of the ongoing write function. */
     const uint8_t *write_buf;
     /** The length of write_buf in bytes. */
-    uint8_t write_length;
+    uint32_t write_length;
     /** The current position in write_buf. */
-    volatile uint8_t write_pos;
+    volatile uint32_t write_pos;
 };
 
 static pbdrv_uart_dev_t uart_devs[PBDRV_CONFIG_UART_STM32F4_LL_IRQ_NUM_UART];
@@ -66,7 +66,7 @@ pbio_error_t pbdrv_uart_get_instance(uint8_t id, pbdrv_uart_dev_t **uart_dev) {
     return PBIO_SUCCESS;
 }
 
-pbio_error_t pbdrv_uart_read(pbio_os_state_t *state, pbdrv_uart_dev_t *uart, uint8_t *msg, uint8_t length, uint32_t timeout) {
+pbio_error_t pbdrv_uart_read(pbio_os_state_t *state, pbdrv_uart_dev_t *uart, uint8_t *msg, uint32_t length, uint32_t timeout) {
 
     PBIO_OS_ASYNC_BEGIN(state);
 
@@ -108,7 +108,7 @@ pbio_error_t pbdrv_uart_read(pbio_os_state_t *state, pbdrv_uart_dev_t *uart, uin
     PBIO_OS_ASYNC_END(PBIO_SUCCESS);
 }
 
-pbio_error_t pbdrv_uart_write(pbio_os_state_t *state, pbdrv_uart_dev_t *uart, const uint8_t *msg, uint8_t length, uint32_t timeout) {
+pbio_error_t pbdrv_uart_write(pbio_os_state_t *state, pbdrv_uart_dev_t *uart, const uint8_t *msg, uint32_t length, uint32_t timeout) {
 
     PBIO_OS_ASYNC_BEGIN(state);
